Add Zagreb Index and Wiener Index to Mordred (#555)

+++ skfp/fingerprints/_new_mordred/utils/graph_matrix.py
@@ -58,7 +58,7 @@ class AdjacencyMatrix:
         bond counts as one edge. In that case, atom degree equals the atom
         valence, i.e. the number of bonds each atom forms.
         """
-        return self._base.sum(axis=0)
+        return self._base.sum(axis=0, dtype=float)
 
 
 class DistanceMatrix3D:
